@@ -10,8 +10,9 @@
     "logs": "firebase functions:log"
   },
   "engines": {
-    "node": "18"
+    "node": "20"
   },
+  "type": "module",
   "main": "index.js",
   "dependencies": {
     "firebase-admin": "^12.0.0",
