@@ -1,4 +1,6 @@
-import React, { useEffect, useMemo, useState } from 'react';
+import React, { useEffect, useMemo, useRef, useState } from 'react';
+
+import ConfirmModal from '../../../components/ui/ConfirmModal';
 
 const CRC_DENOMINATIONS: readonly number[] = [20000, 10000, 5000, 2000, 1000, 500, 100, 50, 25];
 const USD_DENOMINATIONS: readonly number[] = [100, 50, 20, 10, 5, 1];
@@ -53,11 +55,17 @@ const DailyClosingModal: React.FC<DailyClosingModalProps> = ({
     currentBalanceUSD,
     managerReadonly = false,
 }) => {
+    const modalRef = useRef<HTMLDivElement | null>(null);
+    const managerFieldRef = useRef<HTMLSelectElement | HTMLInputElement | null>(null);
+
     const [closingDateISO, setClosingDateISO] = useState(() => new Date().toISOString());
     const [manager, setManager] = useState('');
     const [notes, setNotes] = useState('');
     const [crcCounts, setCrcCounts] = useState<CountState>(() => buildInitialCounts(CRC_DENOMINATIONS));
     const [usdCounts, setUsdCounts] = useState<CountState>(() => buildInitialCounts(USD_DENOMINATIONS));
+
+    const [confirmDiffOpen, setConfirmDiffOpen] = useState(false);
+    const [pendingSubmitValues, setPendingSubmitValues] = useState<DailyClosingFormValues | null>(null);
 
     const crcFormatter = useMemo(
         () => new Intl.NumberFormat('es-CR', { minimumFractionDigits: 0, maximumFractionDigits: 0 }),
@@ -90,12 +98,43 @@ const DailyClosingModal: React.FC<DailyClosingModalProps> = ({
     const diffUSD = totalUSD - Math.trunc(currentBalanceUSD);
     const hasAnyCash = totalCRC > 0 || totalUSD > 0;
     const submitDisabled = manager.trim().length === 0 || !hasAnyCash;
+    const hasDifferences = diffCRC !== 0 || diffUSD !== 0;
+
+    const submitDisabledReason = useMemo(() => {
+        if (manager.trim().length === 0) {
+            return 'Selecciona un encargado para poder guardar.';
+        }
+        if (!hasAnyCash) {
+            return 'No se puede guardar: el efectivo está en 0. Ingresa el conteo en colones o dólares para realizar el cierre.';
+        }
+        return '';
+    }, [manager, hasAnyCash]);
 
     const differenceLabel = (currency: 'CRC' | 'USD', diff: number) => {
         if (diff === 0) return 'sin diferencias';
         const sign = diff > 0 ? '+' : '-';
         return `${sign} ${formatCurrency(currency, Math.abs(diff))}`;
     };
+
+    const differencesConfirmMessage = useMemo(() => {
+        if (!hasDifferences) return '';
+
+        const lines: string[] = ['Hay diferencias entre el efectivo contado y el saldo registrado.', ''];
+
+        if (diffCRC !== 0) {
+            lines.push(
+                `Colones: contado ${formatCurrency('CRC', totalCRC)} · registrado ${formatCurrency('CRC', currentBalanceCRC)} · diferencia ${differenceLabel('CRC', diffCRC)}`,
+            );
+        }
+        if (diffUSD !== 0) {
+            lines.push(
+                `Dólares: contado ${formatCurrency('USD', totalUSD)} · registrado ${formatCurrency('USD', currentBalanceUSD)} · diferencia ${differenceLabel('USD', diffUSD)}`,
+            );
+        }
+
+        lines.push('', '¿Deseas guardar el cierre de todos modos?');
+        return lines.join('\n');
+    }, [hasDifferences, totalCRC, totalUSD, currentBalanceCRC, currentBalanceUSD, diffCRC, diffUSD]);
 
     useEffect(() => {
         if (!open) return;
@@ -197,7 +236,32 @@ const DailyClosingModal: React.FC<DailyClosingModalProps> = ({
         }
     };
 
+    const focusAdjacentCashInput = (current: HTMLInputElement, direction: 1 | -1) => {
+        const root = modalRef.current;
+        if (!root) return;
+
+        const cashInputs = Array.from(root.querySelectorAll<HTMLInputElement>('input[data-cash-count-input="true"]'));
+        const currentIndex = cashInputs.indexOf(current);
+        if (currentIndex === -1) return;
+
+        const next = cashInputs[currentIndex + direction];
+        if (next) {
+            next.focus();
+            next.select();
+            return;
+        }
+
+        if (direction === 1) {
+            managerFieldRef.current?.focus();
+        }
+    };
+
     const handleCountKeyDown = (event: React.KeyboardEvent<HTMLInputElement>, currency: 'CRC' | 'USD', denom: number) => {
+        if (event.key === 'Enter') {
+            event.preventDefault();
+            focusAdjacentCashInput(event.currentTarget, event.shiftKey ? -1 : 1);
+            return;
+        }
         if (event.key === 'ArrowUp') {
             event.preventDefault();
             incrementCount(currency, denom);
@@ -218,7 +282,7 @@ const DailyClosingModal: React.FC<DailyClosingModalProps> = ({
         const trimmedManager = manager.trim();
         if (!trimmedManager || !hasAnyCash) return;
 
-        onConfirm({
+        const values: DailyClosingFormValues = {
             closingDate: closingDateISO,
             manager: trimmedManager,
             notes,
@@ -226,7 +290,30 @@ const DailyClosingModal: React.FC<DailyClosingModalProps> = ({
             totalUSD,
             breakdownCRC: buildBreakdown(crcCounts, CRC_DENOMINATIONS),
             breakdownUSD: buildBreakdown(usdCounts, USD_DENOMINATIONS),
-        });
+        };
+
+        if (hasDifferences) {
+            setPendingSubmitValues(values);
+            setConfirmDiffOpen(true);
+            return;
+        }
+
+        onConfirm(values);
+    };
+
+    const handleConfirmDifferences = () => {
+        if (!pendingSubmitValues) {
+            setConfirmDiffOpen(false);
+            return;
+        }
+        onConfirm(pendingSubmitValues);
+        setConfirmDiffOpen(false);
+        setPendingSubmitValues(null);
+    };
+
+    const handleCancelDifferences = () => {
+        setConfirmDiffOpen(false);
+        setPendingSubmitValues(null);
     };
 
     const handleClearCounts = () => {
@@ -244,6 +331,7 @@ const DailyClosingModal: React.FC<DailyClosingModalProps> = ({
             <div
                 className="w-full max-w-full sm:max-w-3xl rounded border border-[var(--input-border)] bg-[#1f262a] text-white shadow-lg max-h-[80vh] overflow-hidden flex flex-col"
                 onClick={event => event.stopPropagation()}
+                ref={modalRef}
             >
                 <div className="flex items-center justify-between gap-4 p-5 pb-0">
                     <div className="flex-1" />
@@ -280,10 +368,12 @@ const DailyClosingModal: React.FC<DailyClosingModalProps> = ({
                                                     className="w-24 rounded border border-[var(--input-border)] bg-[var(--input-bg)] p-2 pr-8 text-sm text-center"
                                                     inputMode="numeric"
                                                     aria-label={`Cantidad ${denom} colones`}
+                                                    data-cash-count-input="true"
                                                 />
                                                 <div className="absolute right-1 top-1/2 -translate-y-1/2 flex flex-col items-center select-none">
                                                     <button
                                                         type="button"
+                                                        tabIndex={-1}
                                                         onClick={() => incrementCount('CRC', denom)}
                                                         className="w-5 h-4 leading-[10px] rounded-t bg-transparent text-[var(--muted-foreground)] hover:text-[var(--foreground)]"
                                                         aria-label={`Aumentar ${denom}`}
@@ -292,6 +382,7 @@ const DailyClosingModal: React.FC<DailyClosingModalProps> = ({
                                                     </button>
                                                     <button
                                                         type="button"
+                                                        tabIndex={-1}
                                                         onClick={() => decrementCount('CRC', denom)}
                                                         className="w-5 h-4 leading-[10px] rounded-b bg-transparent text-[var(--muted-foreground)] hover:text-[var(--foreground)]"
                                                         aria-label={`Disminuir ${denom}`}
@@ -334,10 +425,12 @@ const DailyClosingModal: React.FC<DailyClosingModalProps> = ({
                                                     className="w-24 rounded border border-[var(--input-border)] bg-[var(--input-bg)] p-2 pr-8 text-sm text-center"
                                                     inputMode="numeric"
                                                     aria-label={`Cantidad ${denom} dólares`}
+                                                    data-cash-count-input="true"
                                                 />
                                                 <div className="absolute right-1 top-1/2 -translate-y-1/2 flex flex-col items-center select-none">
                                                     <button
                                                         type="button"
+                                                        tabIndex={-1}
                                                         onClick={() => incrementCount('USD', denom)}
                                                         className="w-5 h-4 leading-[10px] rounded-t bg-transparent text-[var(--muted-foreground)] hover:text-[var(--foreground)]"
                                                         aria-label={`Aumentar ${denom}`}
@@ -346,6 +439,7 @@ const DailyClosingModal: React.FC<DailyClosingModalProps> = ({
                                                     </button>
                                                     <button
                                                         type="button"
+                                                        tabIndex={-1}
                                                         onClick={() => decrementCount('USD', denom)}
                                                         className="w-5 h-4 leading-[10px] rounded-b bg-transparent text-[var(--muted-foreground)] hover:text-[var(--foreground)]"
                                                         aria-label={`Disminuir ${denom}`}
@@ -389,6 +483,9 @@ const DailyClosingModal: React.FC<DailyClosingModalProps> = ({
                                     onChange={event => setManager(event.target.value)}
                                     className="rounded border border-[var(--input-border)] bg-[var(--input-bg)] p-2 text-sm"
                                     disabled={loadingEmployees || managerReadonly}
+                                    ref={el => {
+                                        managerFieldRef.current = el;
+                                    }}
                                 >
                                     <option value="">Seleccionar encargado</option>
                                     {employees.map(name => (
@@ -404,6 +501,9 @@ const DailyClosingModal: React.FC<DailyClosingModalProps> = ({
                                     className="rounded border border-[var(--input-border)] bg-[var(--input-bg)] p-2 text-sm"
                                     placeholder="Nombre del encargado"
                                     readOnly={managerReadonly}
+                                    ref={el => {
+                                        managerFieldRef.current = el;
+                                    }}
                                 />
                             )}
                         </div>
@@ -450,17 +550,38 @@ const DailyClosingModal: React.FC<DailyClosingModalProps> = ({
                         >
                             Cancelar
                         </button>
-                        <button
-                            type="button"
-                            onClick={handleSubmit}
-                            disabled={submitDisabled}
-                            className="rounded bg-blue-500 px-4 py-2 text-sm font-medium text-white hover:bg-blue-600 disabled:opacity-50"
-                        >
-                            {editId ? 'Actualizar cierre' : 'Crear cierre'}
-                        </button>
+                        <div className="relative group">
+                            <button
+                                type="button"
+                                onClick={handleSubmit}
+                                disabled={submitDisabled}
+                                className="rounded bg-blue-500 px-4 py-2 text-sm font-medium text-white hover:bg-blue-600 disabled:opacity-50"
+                            >
+                                {editId ? 'Actualizar cierre' : 'Guardar cierre'}
+                            </button>
+                            {submitDisabled && submitDisabledReason ? (
+                                <div
+                                    className="pointer-events-none absolute bottom-full right-0 mb-2 w-72 rounded border border-yellow-300 bg-yellow-200 px-3 py-2 text-xs text-black opacity-0 shadow-lg transition-opacity group-hover:opacity-100 dark:border-yellow-600 dark:bg-yellow-500"
+                                    role="tooltip"
+                                >
+                                    {submitDisabledReason}
+                                </div>
+                            ) : null}
+                        </div>
                     </div>
                 </div>
             </div>
+
+            <ConfirmModal
+                open={confirmDiffOpen}
+                title="Confirmar cierre con diferencias"
+                message={differencesConfirmMessage}
+                confirmText={editId ? 'Actualizar de todos modos' : 'Guardar de todos modos'}
+                cancelText="Revisar"
+                actionType="change"
+                onConfirm={handleConfirmDifferences}
+                onCancel={handleCancelDifferences}
+            />
         </div>
     );
 };
