@@ -32,7 +32,7 @@ export function useProviders(company?: string) {
 		}
 	}, [company]);
 
-	const addProvider = useCallback(async (name: string, type?: string, correonotifi?: string) => {
+	const addProvider = useCallback(async (name: string, type?: string, correonotifi?: string, visit?: ProviderEntry['visit']) => {
 		const trimmedCompany = (company || '').trim();
 		if (!trimmedCompany) {
 			const message = 'No se pudo determinar la empresa del usuario.';
@@ -42,7 +42,7 @@ export function useProviders(company?: string) {
 
 		try {
 			setError(null);
-			await ProvidersService.addProvider(trimmedCompany, name, type, correonotifi);
+			await ProvidersService.addProvider(trimmedCompany, name, type, correonotifi, visit);
 			await fetchProviders();
 		} catch (err) {
 			const message = err instanceof Error ? err.message : 'No se pudo guardar el proveedor.';
@@ -72,7 +72,7 @@ export function useProviders(company?: string) {
 		}
 	}, [company, fetchProviders]);
 
-	const updateProvider = useCallback(async (code: string, name: string, type?: string, correonotifi?: string) => {
+	const updateProvider = useCallback(async (code: string, name: string, type?: string, correonotifi?: string, visit?: ProviderEntry['visit']) => {
 		const trimmedCompany = (company || '').trim();
 		if (!trimmedCompany) {
 			const message = 'No se pudo determinar la empresa del usuario.';
@@ -82,7 +82,7 @@ export function useProviders(company?: string) {
 
 		try {
 			setError(null);
-			await ProvidersService.updateProvider(trimmedCompany, code, name, type, correonotifi);
+			await ProvidersService.updateProvider(trimmedCompany, code, name, type, correonotifi, visit);
 			await fetchProviders();
 		} catch (err) {
 			const message = err instanceof Error ? err.message : 'No se pudo actualizar el proveedor.';
