@@ -4,10 +4,28 @@ import { UsersService } from './users';
 
 export class EmpresasService {
     private static readonly COLLECTION_NAME = 'empresas';
+    private static readonly CACHE_TTL_MS = 30_000;
+    private static empresasCache: { expiresAt: number; data: Empresas[] } | null = null;
+
+    private static cloneEmpresas(list: Empresas[]): Empresas[] {
+        return (list || []).map((e) => ({
+            ...e,
+            empleados: Array.isArray((e as any).empleados)
+                ? (e as any).empleados.map((emp: any) => ({ ...emp }))
+                : [],
+        }));
+    }
 
     // Normalize empleados payload to a consistent shape before persisting
     private static normalizeEmpleado(raw: unknown): EmpresaEmpleado {
-        const defaultEmp: EmpresaEmpleado = { Empleado: '', hoursPerShift: 8, extraAmount: 0, ccssType: 'TC' };
+        const defaultEmp: EmpresaEmpleado = {
+            Empleado: '',
+            hoursPerShift: 8,
+            extraAmount: 0,
+            ccssType: 'TC',
+            calculoprecios: false,
+            amboshorarios: false
+        };
         if (!raw || typeof raw !== 'object') return defaultEmp;
 
         const obj = raw as Record<string, unknown>;
@@ -29,6 +47,21 @@ export class EmpresasService {
                 if (v !== undefined && v !== null && v !== '') {
                     const n = Number(v as unknown as number);
                     if (!Number.isNaN(n)) return n;
+                }
+            }
+            return undefined;
+        };
+
+        const getBoolean = (...keys: string[]) => {
+            for (const k of keys) {
+                const v = obj[k];
+                if (v === undefined || v === null) continue;
+                if (typeof v === 'boolean') return v;
+                if (typeof v === 'number') return v !== 0;
+                if (typeof v === 'string') {
+                    const s = v.trim().toLowerCase();
+                    if (s === 'true' || s === '1' || s === 'si' || s === 'sí' || s === 'yes') return true;
+                    if (s === 'false' || s === '0' || s === 'no') return false;
                 }
             }
             return undefined;
@@ -56,7 +89,9 @@ export class EmpresasService {
             Empleado: name || '',
             hoursPerShift: typeof hours === 'number' ? hours : 8,
             extraAmount: typeof extra === 'number' ? extra : 0,
-            ccssType: (ccss === 'MT' || ccss === 'TC') ? (ccss as 'TC' | 'MT') : 'TC'
+            ccssType: (ccss === 'MT' || ccss === 'TC') ? (ccss as 'TC' | 'MT') : 'TC',
+            calculoprecios: getBoolean('calculoprecios', 'calculoPrecios', 'calculo_precios', 'calculo precios') ?? false,
+            amboshorarios: getBoolean('amboshorarios', 'ambosHorarios', 'ambos_horarios', 'ambos horarios') ?? false
         };
 
         return normalized;
@@ -68,17 +103,26 @@ export class EmpresasService {
     }
 
     static async getAllEmpresas(): Promise<Empresas[]> {
+        const cached = this.empresasCache;
+        if (cached && cached.expiresAt > Date.now()) {
+            return this.cloneEmpresas(cached.data);
+        }
+
         const all = await FirestoreService.getAll(this.COLLECTION_NAME) as Empresas[];
-        return all.map(e => ({
+        const normalized = all.map(e => ({
             ...e,
             empleados: EmpresasService.normalizeEmpleados(e.empleados as unknown)
         }));
+        this.empresasCache = { expiresAt: Date.now() + this.CACHE_TTL_MS, data: normalized };
+        return this.cloneEmpresas(normalized);
     }
 
     /**
      * Add a new empresa. If empresa.id is provided, create with that id.
      */
     static async addEmpresa(empresa: Partial<Empresas> & { id?: string }): Promise<string> {
+        // invalidate cache on write
+        this.empresasCache = null;
         // If an ownerId is provided, enforce owner's maxCompanies limit (if any)
         const ownerId = empresa.ownerId || '';
         if (ownerId) {
@@ -122,6 +166,7 @@ export class EmpresasService {
     }
 
     static async updateEmpresa(id: string, empresa: Partial<Empresas>): Promise<void> {
+		this.empresasCache = null;
         const patch = { ...empresa } as Partial<Empresas> & Record<string, unknown>;
         if (patch.empleados) {
             patch.empleados = EmpresasService.normalizeEmpleados(patch.empleados as unknown);
@@ -130,6 +175,7 @@ export class EmpresasService {
     }
 
     static async deleteEmpresa(id: string): Promise<void> {
+		this.empresasCache = null;
         return await FirestoreService.delete(this.COLLECTION_NAME, id);
     }
 }
